+++ nl-linux/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Generate sub-messages.
+
 ## v0.3.0 (2026-09-01)
 
 ## v0.2.4 (2026-08-22)